Program re-write
